Implementing simple SDE with multiple jumps
This is related to https://github.com/jlperla/KnowledgeDiffusionSimulations.jl/issues/1. 
In this julia notebook, we implemented simple SDE with multiple jumps by defining  a vector of jumps. In order to use a vector in `JumpProblem()`, we used `JumpSet((), jumps, nothing, nothing)` which takes `jumps` as an argument. 

In order to solve these sorts of problems with multiple jumps, I compared default solver, `SSAStepper()`, and `SRIW1()`. It seems that even though SSAStepper() is very fast, it interferes with stochastic part of SDE and sort of eliminates it. 

At the end, we implemented a simple model of jump diffusion in which firms adapt new innovations of other firms if they benefits from them. We decided for now to use `SRIW1()`.